Name the hook for DataViews, and keep it beside the modal it feeds
