+++ client/dashboard/components/dataviews/index.tsx
@@ -3,4 +3,3 @@ export * from './dataviews-action-modal';
 export * from './dataviews-card';
 export * from './dataviews-empty-state';
 export * from './dataviews-empty-state-layout';
-export * from './use-deep-linked-action';
